fix: issues from mypy
